move worst_level back into scriptworker.task
