+++ scriptworker/task.py
@@ -26,6 +26,19 @@ from scriptworker.utils import filepaths_in_dir, raise_future_exceptions, retry_
 log = logging.getLogger(__name__)
 
 
+def worst_level(level1, level2):
+    """Given two int levels, return the larger.
+
+    Args:
+        level1 (int): exit code 1.
+        level2 (int): exit code 2.
+
+    Returns:
+        int: the larger of the two levels.
+    """
+    return level1 if level1 > level2 else level2
+
+
 async def run_task(context):
     """Run the task, sending stdout+stderr to files.
